feat: data validations

+++ Controllers/StockContoller.cs
@@ -25,7 +25,7 @@ namespace api.Controllers{
       return Ok(stocks);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id) {
       var stock = await _stockRepo.GetByIdAsync(id);
       if (stock == null) {
@@ -36,14 +36,21 @@ namespace api.Controllers{
 
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto) {
+      if(!ModelState.IsValid){
+       return BadRequest(ModelState);
+      }
      var stockModel = stockDto.ToStockFromCreateDTO();
      await _stockRepo.CreateAsync(stockModel);
      return CreatedAtAction(nameof(GetById), new {id = stockModel.Id}, stockModel.ToStockDto());
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto){
+
+      if(!ModelState.IsValid){
+       return BadRequest(ModelState);
+      }
       var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
       if (stockModel == null) {
         return NotFound();
@@ -52,8 +59,12 @@ namespace api.Controllers{
     }	
 
     [HttpDelete]	
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id){
+      if(!ModelState.IsValid){
+       return BadRequest(ModelState);
+      }
+
       var stockModel  = await _stockRepo.DeleteAsync(id);
       
       if (stockModel == null) {
